multi-rag-commerce-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.1
chromadb==0.4.18
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
tiktoken==0.5.1
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-multipart==0.0.6 